refactor(gate-manager): decompose review_gate_manager.py into 4 mixin modules
Split 1017L monolithic file into focused modules:
- gate_executor.py (141L): gate execution and status queries
- gate_request_handler.py (376L): request creation for Gemini/Codex/Claude GitHub
- gate_result_parser.py (203L): finding classification and result recording
- gate_report_generator.py (138L): audit trail and result persistence
- review_gate_manager.py (254L): facade with init, path helpers, CLI, mixin composition

All modules under 400L limit. Mixin-based class composition preserves
ReviewGateManager interface. 90/91 tests pass (1 pre-existing failure).

Dispatch-ID: 20260405-114646-f27-pr3-gate-manager-decompose-C
